add source name in diagnostics display, apply suggestions
